netlify/functions/requirements.txt
pinecone-client==3.0.3
groq==0.4.1
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.1.0
transformers==4.35.0